指摘修正３
